// studio-sdk/types/contracts/MenuTypes.ts
/**
 * MenuItems are POJOs, generated by MenuItem factory functions registered via `MenuMediator`.
 */
export declare type MenuFactoryFunction = (...args: any[]) => Menu | Promise<Menu>;
export declare type Menu = MenuItem[];
export declare type MenuItem = MenuItem_SubMenu | MenuItem_Command | MenuItem_Divider | MenuItem_Role;
export declare type MenuItem_SubMenu = {
  type: 'menu';
  readonly id?: string;
  /**
   * Defaults to "true".
   */
  readonly visible?: boolean;
  /**
   * Text to be displayed
   */
  readonly label: string;
  readonly submenu: Menu;
};
export declare type MenuItem_Command = {
  type: 'command';
  readonly id?: string;
  /**
   * Defaults to "true".
   */
  readonly visible?: boolean;
  /**
   * Text to be displayed
   */
  readonly label: string;
  /**
   * Whether or not a check mark should be displayed
   */
  readonly checked?: boolean;
  /**
   * The icon id for the icon shown in the menu (if it is rendered in HTML)
   */
  readonly icon?: string;
  /**
   * Name of the command to execute on click
   */
  readonly command: string;
  /**
   * Optional: Arguments for the command
   */
  readonly commandArgs?: any[];
  /**
   * Set to a formatted, human-readable version of the keystroke activating the `command`
   */
  formattedKeystroke?: string;
};
export declare type MenuItem_Divider = {
  type: 'divider';
  readonly id?: string;
  /**
   * Defaults to "true".
   */
  readonly visible?: boolean;
};
/**
 * Used in the electron app to call platform-specific Menu features
 */
export declare type MenuItem_Role = {
  type: 'role';
  readonly id?: string;
  /**
   * Defaults to "true".
   */
  readonly visible?: boolean;
  readonly label?: string;
  readonly role: string;
};